+++ authz/build.gradle
@@ -1,10 +1,10 @@
 plugins {
     id 'com.flipkart.varadhi.java-library-conventions'
+    id 'java-test-fixtures'
 }
 
 dependencies {
     implementation(project(':common'))
-    implementation(project(':entities'))
 
     implementation("io.vertx:vertx-core")
     implementation("org.apache.logging.log4j:log4j-slf4j2-impl")
@@ -14,5 +14,4 @@ dependencies {
 
     testImplementation("io.vertx:vertx-junit5")
     testImplementation("io.vertx:vertx-web-client")
-    testImplementation(testFixtures(project(':entities')))
 }
