+++ authz/src/main/java/com/flipkart/varadhi/config/AuthorizationOptions.java
@@ -1,4 +1,4 @@
-package com.flipkart.varadhi.authz;
+package com.flipkart.varadhi.config;
 
 import lombok.Data;
