+++ authz/src/main/java/com/flipkart/varadhi/entities/ResourceType.java
@@ -1,4 +1,4 @@
-package com.flipkart.varadhi.auth;
+package com.flipkart.varadhi.entities;
 
 public enum ResourceType {
     ORG("org"),
