+++ authz/src/test/java/com/flipkart/varadhi/authz/DefaultAuthorizationProviderTest.java
@@ -1,7 +1,8 @@
 package com.flipkart.varadhi.authz;
 
+import com.flipkart.varadhi.config.AuthorizationOptions;
 import com.flipkart.varadhi.exceptions.InvalidConfigException;
-import com.flipkart.varadhi.auth.ResourceAction;
+import com.flipkart.varadhi.entities.ResourceAction;
 import io.vertx.junit5.Checkpoint;
 import io.vertx.junit5.VertxExtension;
 import io.vertx.junit5.VertxTestContext;
