+++ server/build.gradle
@@ -51,7 +51,7 @@ dependencies {
 
     testImplementation(project(":pulsar"))
     testImplementation(testFixtures(project(":core")))
-    testImplementation(testFixtures(project(":entities")))
+    testImplementation(testFixtures(project(":authz")))
     testImplementation("org.apache.curator:curator-test")
     testImplementation("io.vertx:vertx-junit5")
     testImplementation("io.vertx:vertx-web-client")
