+++ server/src/main/java/com/flipkart/varadhi/auth/PermissionAuthorization.java
@@ -1,5 +1,7 @@
 package com.flipkart.varadhi.auth;
 
+import com.flipkart.varadhi.entities.ResourceAction;
+
 public record PermissionAuthorization(ResourceAction action, ResourceName resource) {
 
     public static PermissionAuthorization of(ResourceAction action, String resource) {
