+++ server/src/main/java/com/flipkart/varadhi/web/AuthorizationHandlerBuilder.java
@@ -2,7 +2,7 @@ package com.flipkart.varadhi.web;
 
 import com.flipkart.varadhi.authz.AuthorizationProvider;
 import com.flipkart.varadhi.auth.PermissionAuthorization;
-import com.flipkart.varadhi.auth.ResourceAction;
+import com.flipkart.varadhi.entities.ResourceAction;
 import com.flipkart.varadhi.entities.UserContext;
 import com.flipkart.varadhi.entities.VertxUserContext;
 import io.vertx.core.Future;
