+++ server/src/main/java/com/flipkart/varadhi/web/v1/produce/ProduceHandlers.java
@@ -33,7 +33,7 @@ import static com.flipkart.varadhi.Constants.PathParams.REQUEST_PATH_PARAM_TOPIC
 import static com.flipkart.varadhi.MessageConstants.ANONYMOUS_PRODUCE_IDENTITY;
 import static com.flipkart.varadhi.MessageConstants.Headers.*;
 import static com.flipkart.varadhi.MessageConstants.PRODUCE_CHANNEL_HTTP;
-import static com.flipkart.varadhi.auth.ResourceAction.TOPIC_PRODUCE;
+import static com.flipkart.varadhi.entities.ResourceAction.TOPIC_PRODUCE;
 import static com.flipkart.varadhi.web.routes.RouteBehaviour.authenticated;
 import static com.flipkart.varadhi.web.routes.RouteBehaviour.hasBody;
 import static java.net.HttpURLConnection.HTTP_INTERNAL_ERROR;
